cambios orden de tabla y validacion no duplicados

+++ Controladores/forma_pago.php
@@ -34,7 +34,7 @@ switch ($_GET["op"]) {
         $FORMA_DE_PAGO = $body["FORMA_DE_PAGO"];
         $DESCRIPCION = $body["DESCRIPCION"];
         $ESTADO =  $body["ESTADO"];
-        if (verificarExistenciaFpago($FORMA_DE_PAGO) > 0) {
+        if (verificarExistenciaFpago($FORMA_DE_PAGO) > 0 && ! esMismoPago($ID_FPAGO, $FORMA_DE_PAGO)) {
             // Envía una respuesta de conflicto (409) si la Forma de pago ya existe
             http_response_code(409);
             echo json_encode(["error" => "La Forma de Pago ya existe en la base de datos."]);
@@ -60,7 +60,11 @@ switch ($_GET["op"]) {
         $FORMA_DE_PAGO = $body["FORMA_DE_PAGO"];
         $DESCRIPCION = $body["DESCRIPCION"];
         $ESTADO =  $body["ESTADO"];
-
+        if (verificarExistenciaFpago($FORMA_DE_PAGO) > 0) {
+            // Envía una respuesta de conflicto (409) si la Forma de pago ya existe
+            http_response_code(409);
+            echo json_encode(["error" => "La Forma de Pago ya existe en la base de datos."]);
+        } else {
         $date = new DateTime(date("Y-m-d H:i:s"));
         $dateMod = $date->modify("-7 hours");
         $dateNew = $dateMod->format("Y-m-d H:i:s"); 
@@ -93,6 +97,7 @@ switch ($_GET["op"]) {
         {
             $bit->insert_bitacoraModificacion($dateNew, $estadoAntes, $ESTADO, $_SESSION['id_usuario'], 12, "ESTADO", $ID_FPAGO, "MODIFICAR");
         }
+    }
         break;
 
     case "EliminarFpago":
@@ -120,4 +125,23 @@ function verificarExistenciaFpago($FORMA_DE_PAGO) {
     // Devuelve el número de resultados encontrados
     return $row['count'];
 }
+
+function esMismoPago($ID_FPAGO, $FORMA_DE_PAGO) {
+    // Realiza una consulta en la base de datos para verificar si la forma de pago tiene el mismo ID y nombre
+    $sql = "SELECT COUNT(*) AS count FROM tbl_formapago WHERE id_fpago = :id_fpago AND forma_de_pago = :forma_de_pago";
+
+    // Realiza la conexión a la base de datos y ejecuta la consulta
+    $conexion = new Conectar();
+    $conn = $conexion->Conexion();
+    $stmt = $conn->prepare($sql);
+    $stmt->bindParam(':id_fpago', $ID_FPAGO);
+    $stmt->bindParam(':forma_de_pago', $FORMA_DE_PAGO);
+    $stmt->execute();
+    $row = $stmt->fetch(PDO::FETCH_ASSOC);
+
+    // Si el número de resultados encontrados es mayor que 0, significa que la forma de pago tiene el mismo ID y nombre
+    return $row['count'] > 0;
+}
+
+
 ?>
